#121 - Final bug fixes before recording demo video.
BP_HeroInventoryWidget - Set text to Hero1 manually for sake of recording video.
shield - Rename the image asset to fix typo from shiled to shield.
IK_Armor_Data, IK_Passive_Skill_Data - Manually set data. If it caused conflicts when merged, ignore these changes. It is not important one.
LevelTransitionManager - The manager forces us to start game from only the HeroSelectLevel because the manager need to initialize spawn_data_. Thus, the manager crashes the whole game when we start from the HeroSelectLevel. In order to avoid crashes, add a simple if checks. However, it is not an ultimate solution. Basically I need to refactor the LevelTransitionManager in near future.