#123 - Started establishing deck system.
BP_EnemySpawnerManager - Now it has EnemySpawnData to retrieve them freely.

EnemySpawnData - A data asset contains enemy spawn unit data. For sake of test, it now has only compose 1.

CharacterStatComponent - Instead of using assignment operator, call Setter function of CharacterData.

EnemySpawnDataAsset - A data asset that holds EnemySpawnData.
EnemySpawnData - A struct that has a pointer to EnemySpawnUnitAsset.
EnemySpawnUnitAsset - A data asset that contains an array of EnemySpawnUnit.
EnemySpawnUnit - A struct that manages a unit of enemy spawn data. It means a couple of data such as enemy class that will be spawned and a vector that indicates spawn offset.

The reason why I instantiated four class and structs of it is decoupling of data assets. In order to make maintenance easier when the data assets broken.

EnemySpawnerManager - Change old code to adapt new code.

IKPlayerCameraManager, IKPlayerController - Make it takes a raw pointer instead of TWeakObjectPtr because function parameters are not the place where passed data preserved. They are using registers.