#58, #123 - Handles defeat game flow, debug stat buff of HP&Shield.
DefaultEngine - Modify collision project setting to resolve a bug bullets overlaps with UIs.

NS_ElectricsBullets, NS_Viper - Worked to find a new FX for current bullet styles, but failed to find the best for now.

CharacterStatComponent - Refactored a logic to work correctly when a buff came on HP or shield. However, it works only for a percentage buff. Fixed value buff on HP or Shield does not work. It maybe considered as invalid behavior.

IKGameModeBase - Improve a workflow to handle defeat game logic.