#29 - Improve targeting visual feedback.
TargetingComponent - Apply highlight material on not only actor seleting mode, but also location mode and direction mode.

***

As far as I profiled it, UTargetingComponent::Tick function takes 50 nano seconds. I thought it is okay in terms of performance. However, NS_Impact_Sparks takes 11% of all tick threads. It might be the first case if we need to consider performance.