#228 - Improved game flow after players defeated.
CombatLevel - Magnify NavMesh.

LevelTransitionSubsystem - Changed a moment to call a function that updates buff duration because GlobalBuffDisplayer is in MapLevel. It is weird to be visible a buff that will be deleted when player go to next level.

IKGameInstance - Call InitSpawnData instead of ClearSpawnData. Remain ClearSpawnData function because it is called in ULevelTransitionSubsystem::Initialize.