#56, #94, #194, #233 - Display 0 damage again, increase credit rewards, Improve sentry gun VFX.
BP_SentryGun - Move static mesh to be origin of the mesh to be center of the mesh.

NS_DC_Smokes - Change decal size. Modify spawn location of debris.

Unit - Do not skip 0 damage UI.

IKGameModeBase - increase an amount of credit rewards (from 10 to 50). One combat can buy an item in the store level.