#29 - Debugged no target problem.
There was a bug skill casted even though there are no targets in the range. I resolved it by creating an another function named CanActivateSkill to SkillBase class inspires by GameAbilitySystem in Unreal. Unfortunately, support skills are not adapted it yet.

All files - Implement CanActivateSkill describes a condition that the skill is castable.